backend/pyproject.toml
[tool.poetry]
name = "back"
version = "0.1.0"
description = "Monitoring System"
authors = ["Koral AT"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
pydantic = "^2.10.2"
cassandra-driver = "^3.29.2"
pydantic-settings = "^2.6.1"
flask = "^3.1.0"
flask-sqlalchemy = "^3.1.1"
flask-marshmallow = "^1.2.1"
flask-restful = "^0.3.10"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
